Game Version 0.00043 (Minor Update)
Added a battle log to battle for future use (may remove later)
Updated battle factory to track difficulty
Added Health scaler to the enemy
Added the skill damage scaler to difficulty
Fixed exp and check level bug in player

note: Battle log was added for a first version of testing the neural network to control enemy action but may be scrapped as model is not working. Supervised learning not bet approach for this, learning RL to try and implement in later.